refactor [#2] Component 파일 이름을 pascal case로 변경
